// app/registrar/types.ts
//type
export interface Persona {
    nombres: string;
    paterno: string;
    materno: string;
    provincia: string;
    departamento: string;
    localidad: string;
    oficialia?: number;
    libro?: number;
    partida?: number;
    folio?: number;
    f_nacimiento: string;
    ci: string;
    complemento: string;
    expedido: string;
}